Remove circular dependency between Vector3 and Basis.
