+++ core/math/vector3.h
@@ -96,8 +96,8 @@ struct Vector3 {
 
 	_FORCE_INLINE_ Vector3 cross(const Vector3 &p_b) const;
 	_FORCE_INLINE_ real_t dot(const Vector3 &p_b) const;
-	_FORCE_INLINE_ Basis outer(const Vector3 &p_b) const;
-	_FORCE_INLINE_ Basis to_diagonal_matrix() const;
+	Basis outer(const Vector3 &p_b) const;
+	Basis to_diagonal_matrix() const;
 
 	_FORCE_INLINE_ Vector3 abs() const;
 	_FORCE_INLINE_ Vector3 floor() const;
@@ -154,9 +154,6 @@ struct Vector3 {
 	_FORCE_INLINE_ Vector3() { x = y = z = 0; }
 };
 
-// Should be included after class definition, otherwise we get circular refs
-#include "core/math/basis.h"
-
 Vector3 Vector3::cross(const Vector3 &p_b) const {
 
 	Vector3 ret(
@@ -170,21 +167,6 @@ Vector3 Vector3::cross(const Vector3 &p_b) const {
 real_t Vector3::dot(const Vector3 &p_b) const {
 
 	return x * p_b.x + y * p_b.y + z * p_b.z;
-}
-
-Basis Vector3::outer(const Vector3 &p_b) const {
-
-	Vector3 row0(x * p_b.x, x * p_b.y, x * p_b.z);
-	Vector3 row1(y * p_b.x, y * p_b.y, y * p_b.z);
-	Vector3 row2(z * p_b.x, z * p_b.y, z * p_b.z);
-
-	return Basis(row0, row1, row2);
-}
-
-Basis Vector3::to_diagonal_matrix() const {
-	return Basis(x, 0, 0,
-			0, y, 0,
-			0, 0, z);
 }
 
 Vector3 Vector3::abs() const {
